pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kelvin"
version = "0.1.0"
description = "An unsupervised correctness signal for RAG pipelines."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Kelvin contributors" }]
keywords = ["rag", "evaluation", "metamorphic-testing", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pyyaml>=6",
    "jinja2>=3",
]

[project.scripts]
kelvin = "kelvin.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kelvin"]

[tool.hatch.build.targets.wheel.force-include]
"src/kelvin/reporters/templates" = "kelvin/reporters/templates"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = ["I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra"
